requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
android==1.0
anyio==4.9.0
APScheduler==3.11.0
astroid==3.3.9
attrs==25.3.0
audiostream==0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cmake @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/cmake-3.31.5-py3-none-any.whl#sha256=0c372e231235ded65541b768b4450199bf9238b4e3dcdb8885f17c6d1f9007b6
dill==0.3.9
distro==1.9.0
docutils==0.21.2
feedparser==6.0.11
filetype==1.2.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
jedi==0.19.2
Kivy==2.3.1
lazy-object-proxy==0.0.0
mccabe==0.7.0
multidict==6.5.1
nest-asyncio==1.6.0
ninja @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/ninja-1.12.1-py3-none-any.whl#sha256=369aa5ff36a1bbe7299b02cd6d67f2589e7fd3e3871303c13bafa62a87b5b4fc
packaging==24.2
parso==0.8.4
pillow==11.3.0
pkgconfig==1.5.5
platformdirs==4.3.7
propcache==0.3.2
pybind11==2.13.6
pycparser==2.22
pygame==2.6.1
pyjnius==1.6.1
pylint==3.3.4
PySDL2==0.9.17
python-dotenv==1.1.1
python-telegram-bot==22.1
pytz==2025.2
requests==2.32.4
scikit-build==0.18.1
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tomlkit==0.13.2
tornado==6.5.1
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
